+++ PersianCalendar/src/main/java/com/byagowi/persiancalendar/ui/about/DeviceInformationFragment.kt
@@ -263,8 +263,7 @@ fun <T> T.circularRevealFromMiddle() where T : View?, T : CircularRevealWidget {
             AnimatorSet().also {
                 it.playTogether(
                     CircularRevealCompat.createCircularReveal(
-                        this, (viewWidth / 2).toFloat(), (viewHeight / 2).toFloat(),
-                        10f, (viewDiagonal / 2).toFloat()
+                        this, viewWidth / 2f, viewHeight / 2f, 10f, viewDiagonal / 2f
                     ),
                     ObjectAnimator.ofArgb(
                         this,
